+++ app/src/main/java/woowacourse/shopping/ui/order/OrderViewModel.kt
@@ -14,7 +14,7 @@ import woowacourse.shopping.domain.repository.OrderRepository
 import woowacourse.shopping.domain.repository.ProductHistoryRepository
 import woowacourse.shopping.domain.repository.ProductsRecommendationRepository
 import woowacourse.shopping.domain.repository.ShoppingCartRepository
-import woowacourse.shopping.ui.OnItemQuantityChangeListener
+import woowacourse.shopping.ui.order.listener.OrderListener
 import woowacourse.shopping.ui.util.MutableSingleLiveData
 import woowacourse.shopping.ui.util.SingleLiveData
 import woowacourse.shopping.ui.util.UniversalViewModelFactory
