+++ app/src/main/java/woowacourse/shopping/ui/order/listener/OnOrderListener.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.order
+package woowacourse.shopping.ui.order.listener
 
 interface OnOrderListener {
     fun createOrder()
