+++ app/src/main/java/woowacourse/shopping/ui/order/listener/OrderListener.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.ui.order
+package woowacourse.shopping.ui.order.listener
 
 import woowacourse.shopping.ui.OnItemQuantityChangeListener
