+++ app/src/main/res/layout/layout_recommended_order.xml
@@ -7,7 +7,7 @@
 
         <variable
             name="onOrderListener"
-            type="woowacourse.shopping.ui.order.OnOrderListener" />
+            type="woowacourse.shopping.ui.order.listener.OnOrderListener" />
 
         <variable
             name="totalCartItemQuantity"
